backend/requirements.txt
asgiref==3.2.9
certifi==2020.6.20
chardet==3.0.4
coverage==5.1
Django==3.0.7
django-cors-headers==3.4.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
gunicorn==20.0.4
idna==2.9
PyJWT==1.7.1
pytz==2020.1
requests==2.24.0
sqlparse==0.3.1
urllib3==1.25.9
whitenoise==5.1.0
